requirements.gpu.txt
# PyTorch GPU builds (CUDA 12.1)를 위한 추가 인덱스
--extra-index-url https://download.pytorch.org/whl/cu121

# -------------------------------------------------
# --- 핵심 AI & 오디오 라이브러리 (Core AI & Audio) ---
# -------------------------------------------------

# PyTorch 및 관련 라이브러리 (GPU 버전)
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0

# 음악 소스 분리 라이브러리
demucs==4.0.1

# 오디오 처리를 위한 클래식 라이브러리
librosa==0.11.0
# -------------------------------------------------
# --- 벡터 검색 (Vector Search) ---
# -------------------------------------------------

# Faiss (GPU 버전)
faiss-gpu==1.8.0

# -------------------------------------------------
# --- 웹 서버 (Web Server) ---
# -------------------------------------------------
fastapi==0.120.0
uvicorn[standard]==0.35.0

# -------------------------------------------------
# --- 유틸리티 (Utilities) ---
# -------------------------------------------------

numpy==1.26.4

# 진행률 표시줄
tqdm==4.68.0

requests==2.35.0
